Reorganize components and hooks
