+++ apps/drain-safe/package.json
@@ -5,7 +5,6 @@
   "dependencies": {
     "@gnosis.pm/safe-react-components": "^0.6.0",
     "@material-ui/core": "^4.11.0",
-    "exponential-backoff": "^3.1.0",
     "web3-eth-abi": "^1.3.5"
   },
   "scripts": {
