+++ apps/drain-safe/src/utils/fetch-json.ts
@@ -1,7 +1,9 @@
-export async function fetchJson(url: string): Promise<unknown> {
+async function fetchJson(url: string): Promise<unknown> {
   const resp = await fetch(url);
   if (!resp.ok) {
     throw new Error('Network response was not ok');
   }
   return resp.json();
 }
+
+export default fetchJson;
